Adds a filesystem free method
Prior to this commit, S0007 used existing facts to check the free disk
space. This commit adds a shared method to check the disk space of an
arbitrary filesytem path and uses it for S0007 and S0008.

Prior to this commit, the acceptance test was checking that all elements
in the fact hash were not false, however it did not specify the number
of elements in the hash This could lead to a scenario where a chunk is
missing and still passes. This commit checks the explicit number of
chunks in the hash.