(fix) Ensure acceptance test validates all facts
Prior to this commit, the acceptance test was checking that all elements
in the fact hash were not false, however it did not specify the number
of elements in the hash This could lead to a scenario where a chunk is
missing and still passes. This commit checks the explicit number of
chunks in the hash.