(SUP-2892) - No recent OOM errors logged in any JVM Previously, we had no way to check for OOM errors. This change will be able to provide more information as to which service is running out of memory. S0016-18
Fix path Updating the path variable to go back one directory
Flipped the logic include? returns 'false' if the error is not in the logfile causing to throw errors when the OOM are not present.

Change the number of facts Change the number of facts in self_service_spec.rb

PDK Validate corrections