(SUP-2918) Update S0039 to parse server access log
Prior to this commit, we parsed puppetserver.log log check for 503
errors, but this can be determined from the access logs without the risk
of parsing a huge file with many custom code errors and warnings.  This
commit uses an Apache regex to check if puppetserver-access.log has 503
return codes.