(SUP-3101)
This commit modifies S0022
Previously if end date did not exist or resulted in an invalid date,
Facter would experience a runtime exception and crash.
Additionally, the default behavior was for validity to resolve to
true.
Now pe_status_check.S0022 handles the exception when end date is missing or
results in an invalid date. It prints a Facter warning
message, and the fact resolves to false if end date is missing or
cannot be converted to a date. If license_type is
not Perpetual or Subscription or is missing the S0022 resolves to false.

Previously the grep result, an array, for license_type and end date
would be converted into a string. This commit alters this logic to
capture the first item in the array returned from the grep.
Previously: Date.parse(["end: 2022-05-05\n"])
Now: Date.parse("end: 2022-05-05")
Error handling has been added to account for when license_type is
missing, the first item is nil in the array returned from
File.readlines(license_file).grep(%r{license_type:}).

This commit also modifies the spec test for S0022. Previously the spec
test expected failure if the license file doesn't exist on the machine.
Now the spec test adds a new temporary license file with an end date
that has already passed for the failure condition.