(SUP-3180) Rescue a loaderror when checking filesystem
Prior to this commit, if 'sys/filesystem' failed to load in the fact
resolution, 'puppet facts show' would fail with an empty output. This
would cause problems with `puppet infra run` commands as it uses a
different ruby stack than the puppet-agent. This commit catches the
loaderror and causes any checks that use it to be false.