(SUP-3005) Add a PuppetDB query to the plan
Prior to this commit, it was required to specify a target list to the
plan of the nodes to check. This commit changes the default of the
targets parameter to use a PuppetDB lookup to identify the nodes with
the `pe_build` fact as the target list. After this change, it is easier
to run the plan on all infrastructure nodes when bolt is configured to
connect to PuppetDB.