(PUPSUP-86) Update S0015 and AS001
This commit reduces the threshold to flag S0015 and AS001 as false from
90 days to 30 days. This change was made since in PE 2023.4 certificate
   auto renewal is enabled and new certs having a validity of 90 days.
Therefore, this avoids the indicators perpetually being in a false
state. Also this commit fixes a typo changing 'too' to 'to' for
gramatical clarity in agent_status_check.rb. No tests needed to be
changed for these indicators.